feat: reactivity of each

+++ src/prototy.js
@@ -3,9 +3,8 @@ import { isObject } from './utils/isObject'
 import { isEqual } from './utils/isEqual'
 import { createDynamicFunction } from './utils/createDynamicFunction'
 import Directives from './directives/Directives'
-import { AttributeCache } from './reactivity/attributeCache.js'
+import { Reactivity } from './reactivity.js'
 import { addEvent } from './utils/addEvent'
-import { trigger } from './reactivity/trigger'
 
 /**
  * @typedef {object} PrototyOptions
@@ -19,40 +18,32 @@ class Prototy {
 	/**
 	 * @param {PrototyOptions} options
 	 */
-	  constructor(options = {
-		state: {},
-		root: document.body,
-		static: {},
-		handles: {},
-		directives: {}
-	}) {
-	    this.root = options.root
-	    this.directive = new Directives(options.directives, this.setup.bind(this))
+	constructor(options = { state: {}, root: document.body, static: {}, handles: {}, directives: {} }) {
+		this.root = options.root
+		this.static = options.static
+		this.state = this.createProxy(options.state)
+		this.directive = new Directives(options.directives, this.setup.bind(this))
+		this.reactivity = new Reactivity()
 
-	    /** @type {object} */
-	    this._state = options.state
+		/** @type {Record<string, Function>} */
+		this.handles = {}
+		this.pendingPaths = new Set()
+	  this.delayedAddToCache = () => {}
 
-	    /** @type {object} */
-	    this.static = options.static
-	    this.state = this.createProxy(options.state)
-	    /** @type {Record<string, Function>} */
-	    this.handles = {}
-	    this.pendingPaths = new Set()
-		this.delayedAddToCache = () => {}
-
-	    if (options.handles) {
-	      Object.keys(options.handles).forEach((key) => {
-	        if (typeof options.handles[key] === 'function') {
-	          this.handles[key] = options.handles[key].bind(this)
-	        }
-	      })
-	    }
-		this.bus = {
-			state: this.state,
-			static: this.static,
-			handles: this.handles
+		if (options.handles) {
+		  Object.keys(options.handles).forEach((key) => {
+		    if (typeof options.handles[key] === 'function') {
+		      this.handles[key] = options.handles[key].bind(this)
+		    }
+		  })
 		}
-		document.addEventListener('DOMContentLoaded', () => this.setup(this.root))
+
+	  this.bus = {
+		  state: this.state,
+		  static: this.static,
+		  handles: this.handles
+	  }
+	  document.addEventListener('DOMContentLoaded', () => this.setup(this.root))
 	}
 
 	/**
@@ -60,10 +51,6 @@ class Prototy {
 	 * @param {object} item
 	 */
 	setup(node, item) {
-
-		// eslint-disable-next-line sonarjs/constructor-for-side-effects
-		new AttributeCache(node)
-
 		findElements(node, (/** @type {HTMLElement} */  element, /** @type {string} */ key, /** @type {string} */ code) => {
 			const func = createDynamicFunction(code, this.bus, 'item')
 
@@ -71,7 +58,7 @@ class Prototy {
 				const res = func(item)
 				this.directive.apply(element, key, res)
 			}
-			this.delayedAddToCache = (path) => node._cache.add(element, key, path, update.bind(this))
+			this.delayedAddToCache = (path) => this.reactivity.add(element, key, path, update.bind(this))
 
 			this.activeEffect = update
 			update()
@@ -118,45 +105,47 @@ class Prototy {
 
 	        return value
 	      },
-	      set(target, property, value) {
-	        if (typeof property === 'symbol') {
-	          return Reflect.set(target, property, value)
-	        }
-	        /** @type {Record<string | symbol, any>} */
-	        const t = target
-	        const oldValue = Reflect.get(t, property)
-	        const fullPath = path
-	          ? `${path}.${property.toString()}`
-	          : property.toString()
+		    set(target, property, value) {
+			    if (typeof property === 'symbol') {
+				    return Reflect.set(target, property, value)
+			    }
 
-	        let newValue = value
-	        if (isObject(value)) {
-	          newValue = self.createProxy(value, fullPath)
-	        }
-	        const success = Reflect.set(t, property, newValue)
+			    const t = target
+			    const oldValue = Reflect.get(t, property)
+			    const fullPath = path
+				    ? `${path}.${property.toString()}`
+				    : property.toString()
 
-	        if (success && !isEqual(oldValue, newValue)) {
-	          const parts = fullPath.split('.')
-	          if (
-	            Array.isArray(target) &&
-	            (/^\d+$/.test(property) || property === 'length')
-	          ) {
-						if (!self.pendingPaths.has(path)) {
-							self.pendingPaths.add(path)
-							queueMicrotask(() => {
-								trigger(self.root._cache, path)
-								self.pendingPaths.delete(path)
-							})
-						}
-					} else if (parts.length >= 3 && (/^\d+$/.test(parts[parts.length - 2]))) {
-						trigger(self.root._cache, fullPath, parts)
-					} else {
-						trigger(self.root._cache, fullPath)
-					}
-				}
-				return success
-			}
+			    let newValue = value
+			    if (isObject(value)) {
+				    newValue = self.createProxy(value, fullPath)
+			    }
+
+			    const success = Reflect.set(t, property, newValue)
+
+			    if (success && !isEqual(oldValue, newValue)) {
+				    const isArrayIndex = Array.isArray(target) && (/^\d+$/.test(property) || property === 'length')
+
+				    if (isArrayIndex && !self.pendingPaths.has(path)) {
+					    self.pendingPaths.add(path)
+					    queueMicrotask(() => {
+						    self.trigger(fullPath)
+						    self.pendingPaths.delete(path)
+					    })
+				    } else {
+					    self.trigger(fullPath)
+				    }
+			    }
+			    return success
+		    }
 		})
+	}
+	/**
+	 * @param {string|number} path
+	 */
+	trigger(path) {
+		const arr = this.reactivity.find(path)
+		arr.forEach(item => item.update())
 	}
 }
 export default Prototy
